requirements.txt
asgiref==3.8.1
Django==5.0.4
django-render-partial==0.4
jalali_core==1.0.0
jdatetime==5.2.0
path==17.1.0
pillow==11.2.1
pytz==2025.2
sqlparse==0.5.3
asgiref==3.8.1
Django==5.0.4
django-render-partial==0.4
jalali_core==1.0.0
jdatetime==5.2.0
path==17.1.0
pillow==11.2.1
pytz==2025.2
sorl-thumbnail==12.11.0
sqlparse==0.5.3
